+++ lib/flutter_responsive_utils.dart
@@ -9,12 +9,13 @@ library adaptive_design_utils;
 
 import 'package:flutter/material.dart';
 import 'src/screen_utils.dart';
-
 export 'src/screen_utils.dart';
 export 'src/breakpoints.dart';
 export 'src/device_utils.dart';
 export 'src/advanced_layout.dart';
 export 'src/advanced_typography.dart';
+export 'src/responsive_extensions.dart';
+export 'src/alert_utils.dart';
 
 /// The main class for Adaptive Design Utils plugin
 class AdaptiveDesignUtils {